# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["categories"]
      },
      "get": {
        "operationId": "CategoriesController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["categories"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/categories/{categoryId}": {
      "get": {
        "operationId": "CategoriesController_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["categories"]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["categories"]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["categories"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["transactions"]
      },
      "get": {
        "operationId": "TransactionsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["INCOME", "EXPENSE"]
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdByUserId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["transactions"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/transactions/{transactionId}": {
      "get": {
        "operationId": "TransactionsController_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["transactions"]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["transactions"]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["transactions"]
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/recurring-transactions": {
      "post": {
        "operationId": "RecurringTransactionsController_create",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurringTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recurring-transactions"]
      },
      "get": {
        "operationId": "RecurringTransactionsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recurring-transactions"]
      }
    },
    "/api/v1/workspaces/{workspaceId}/recurring-transactions/{id}": {
      "get": {
        "operationId": "RecurringTransactionsController_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recurring-transactions"]
      },
      "patch": {
        "operationId": "RecurringTransactionsController_update",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecurringTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recurring-transactions"]
      },
      "delete": {
        "operationId": "RecurringTransactionsController_remove",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["recurring-transactions"]
      }
    },
...
  },
  "info": {
    "title": "Finance Platform API",
    "description": "MVP REST API for authentication, multi-tenant workspaces, financial operations, budgets, dashboard, and reports.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["INCOME", "EXPENSE"]
          },
          "parentId": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": ["name", "type"]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["INCOME", "EXPENSE"]
          },
          "parentId": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["INCOME", "EXPENSE"]
          },
          "amount": {
            "type": "string",
            "example": "59.90"
          },
          "transactionDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": ["accountId", "categoryId", "type", "amount", "transactionDate", "description"]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["INCOME", "EXPENSE"]
          },
          "amount": {
            "type": "string",
            "example": "59.90"
          },
          "transactionDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "CreateRecurringTransactionDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["INCOME", "EXPENSE"]
          },
          "amount": {
            "type": "string"
          },
          "frequency": {
            "type": "string",
            "enum": ["DAILY", "WEEKLY", "MONTHLY", "YEARLY"]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "nextExecutionDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "accountId",
          "categoryId",
          "type",
          "amount",
          "frequency",
          "startDate",
          "nextExecutionDate",
          "description"
        ]
      },
      "UpdateRecurringTransactionDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["INCOME", "EXPENSE"]
          },
          "amount": {
            "type": "string"
          },
          "frequency": {
            "type": "string",
            "enum": ["DAILY", "WEEKLY", "MONTHLY", "YEARLY"]
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "nextExecutionDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        }
      }
    }
  }
}